+++ misc.lua
@@ -1,7 +1,8 @@
 -- Misc
 
--- Huntress : Set Stealth Hunting upgrade
+-- Set special upgrades
 Skill.find("aphelion-huntressStealth").upgrade_skill = Skill.find("aphelion-huntressStealthBoosted")
+Skill.find("aphelion-sniperBlast").upgrade_skill = Skill.find("aphelion-sniperBlastBoosted")
 
 -- Huntress : Grant 15 i-frames on Blink
 Skill.find("ror-huntressC"):onActivate(function(actor, skill, index)