+++ misc.lua
@@ -1,9 +1,10 @@
 -- Misc
 
--- Huntress : Set Stealth Hunting upgrade
+-- Set special upgrades
 Skill.find("aphelion-huntressStealth").upgrade_skill = Skill.find("aphelion-huntressStealthBoosted")
+Skill.find("aphelion-sniperBlast").upgrade_skill = Skill.find("aphelion-sniperBlastBoosted")
 
 -- Huntress : Grant 15 i-frames on Blink
-Player:onSkillUse("aphelion-huntressBlinkTweak", function(actor)
+Skill.find("ror-huntressC"):onActivate(function(actor, skill, index)
     actor:set_immune(15)
-end, Skill.find("ror-huntressC"))
+end)